move ConnectionMode's package
