+++ shardingsphere-underlying/shardingsphere-executor/src/main/java/org/apache/shardingsphere/underlying/executor/sql/jdbc/connection/ExecutionConnection.java
@@ -17,6 +17,7 @@
 
 package org.apache.shardingsphere.underlying.executor.sql.jdbc.connection;
 
+import org.apache.shardingsphere.underlying.executor.sql.ConnectionMode;
 import org.apache.shardingsphere.underlying.executor.sql.jdbc.group.StatementOption;
 
 import java.sql.Connection;
