+++ shardingsphere-underlying/shardingsphere-executor/src/test/java/org/apache/shardingsphere/underlying/executor/group/StatementExecuteGroupEngineTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.underlying.executor.group;
 
-import org.apache.shardingsphere.underlying.executor.sql.jdbc.connection.ConnectionMode;
+import org.apache.shardingsphere.underlying.executor.sql.ConnectionMode;
 import org.apache.shardingsphere.underlying.executor.context.ExecutionUnit;
 import org.apache.shardingsphere.underlying.executor.context.SQLUnit;
 import org.apache.shardingsphere.underlying.executor.sql.jdbc.StatementExecuteUnit;
